refactor: rename package to directdefund
